src/requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
blessings==1.7
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
curtsies==0.3.0
defusedxml==0.6.0
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
facebook-sdk==3.1.0
drf-yasg==1.16.0
greenlet==0.4.15
gunicorn==19.4.5
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
oauth2-provider==0.0
oauthlib==3.0.1
Pygments==2.4.2
PyJWT==1.7.1
python-dotenv==0.10.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.21.0
toml==0.10.0
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
